Split functions to separated file for better readability
